Renamed UI test and added Unit Test
