+++ app/src/main/java/io/github/cmput301f19t19/legendary_fiesta/ui/AddPostFragment.java
@@ -363,6 +363,7 @@ public class AddPostFragment extends Fragment implements View.OnClickListener,
      * @param moodId moodId of the mood selected
      * @return Return the id of mood in emotion radio group
      */
+    // TODO: TEST
     private int getEmotionRadioId(@Mood.MoodType int moodId) {
         switch (moodId) {
             case Mood.NEUTRAL:
@@ -386,6 +387,7 @@ public class AddPostFragment extends Fragment implements View.OnClickListener,
     /**
      * @return Return the mood choosen by the RadioGroup
      */
+    // TODO: TEST
     private Mood getSelectedMood(int id) {
         switch (id) {
             case R.id.icon_neutral:
@@ -414,6 +416,7 @@ public class AddPostFragment extends Fragment implements View.OnClickListener,
      * @return
      *  Returns an integer that corresponds to the selected social condition
      */
+    // TODO: TEST
     private Integer getSelectedSocialCondition(String socialCondition) {
         switch (socialCondition) {
             case "Single":
